backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
httpx==0.27.2
beautifulsoup4==4.12.3
lxml==5.3.0
feedparser==6.0.11
python-dateutil==2.9.0
pydantic==2.10.1
playwright==1.48.0
